refactor: renamed `error_handlers.py`
