+++ src/bot/handlers/candidate_commands.py
@@ -17,7 +17,7 @@ from src.database.models import (
 
 from src.bot.utils.bot_answers_json_builder import build_json
 from src.bot.utils.schedule_form_reminder import schedule_form_reminder
-from src.bot.utils.handle_error import handle_db_error
+from src.bot.utils.error_handlers import handle_db_error
 
 from src.database.models.application import ApplicationStatus
 from src.database.models.bot_interaction import InteractionState
